GeometryCore: Image.  Fixing type promotion warnings, and adding a checkslow to help identify a possible overflow case.
#preflight 62bb106d1879a286b893cb66
#rb Jimmy.Andrews

[CL 20879220 by David Hill in ue5-main branch]

+++ Engine/Source/Runtime/GeometryCore/Public/Image/ImageTile.h
@@ -87,7 +87,7 @@ public:
 	/** @return the source image's coordinates given linear index into this tile. */
 	FVector2i GetSourceCoords(const int64 LinearIdx) const
 	{
-		return FVector2i(Start.X + LinearIdx % Width, Start.Y + LinearIdx / Width);
+		return FVector2i(Start.X + int32(LinearIdx % Width), Start.Y + int32(LinearIdx / Width));
 	}
 
 	/** @return the source image's coordinates given local XY coordinates into this tile. */
